remove unnecessary reducer file
