+++ src/redux/configureStore.js
@@ -3,11 +3,9 @@ import { configureStore } from '@reduxjs/toolkit';
 import { combineReducers } from 'redux';
 import thunk from 'redux-thunk';
 import aeroplanesReducer from './aeroplanes/aeroplanes';
-import deleteplanesReducer from './aeroplanes/deleteplane';
 
 const rootReducer = combineReducers({
   aeroplanes: aeroplanesReducer,
-  deleteplanes: deleteplanesReducer
   /* add reducers here */
 });
